requirements.txt
Django==2.1.7
django-countries==5.3.3
django-debug-toolbar==1.11
django-filter==2.1.0
djangorestframework==3.9.2
djangorestframework-simplejwt==4.1.5
Markdown==3.1
Pillow==5.4.1
pkg-resources==0.0.0
psycopg2==2.7.7
psycopg2-binary==2.7.7
PyJWT==1.7.1
pytz==2018.9
sqlparse==0.3.0
